convert all timestamp responses to double precision.

+++ lib/sql/message_room/get.sql
@@ -16,8 +16,8 @@ SELECT 'message_room' AS type,
        (SELECT COUNT(*) FROM messages WHERE message_room = message_rooms.id AND video_url <> NULL)::INT AS video_count,
        (SELECT COUNT(*) FROM messages WHERE message_room = message_rooms.id)::INT AS message_count,
        (SELECT id FROM messages WHERE message_room = message_rooms.id ORDER BY messages.created_at DESC LIMIT 1) AS latest_message,
-       EXTRACT(EPOCH FROM created_at)::INT AS created_at,
-       EXTRACT(EPOCH FROM updated_at)::INT AS updated_at
+       EXTRACT(EPOCH FROM created_at) AS created_at,
+       EXTRACT(EPOCH FROM updated_at) AS updated_at
 FROM message_rooms
 LEFT JOIN invs ON message_rooms.id = invs.message_room
 WHERE id = $1
